+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 This document contains the Spec2nii release history in reverse chronological order.
 
+0.6.1 (Wednesday 18 January 2023)
+---------------------------------
+- Fixed conjugation issue introduced by new nifti-mrs package dependency
+- SPAR/SDAT pipeline now handles HYPER special case.
+- Data/list pipeline now handles HYPER special case.
+- Fixed issue with XA Twix PatientSex and TxOffset attributes.
+- Reenable Bruker conversion. 
+
+0.6.0 (Wednesday 11th January 2023)
+-----------------------------------
+- NIfTI-MRS creation/handling/verification now performed by nifti-mrs package.
+
 0.5.0 (Friday 2nd December 2022)
 --------------------------------
 - NIfTI-MRS version 0.6